Lista de tarefas agora é um componente possibilitando criar listas em diferentes telas ou diferentes contextos.
